+++ src/load_inputs/generate_system.jl
@@ -40,7 +40,15 @@ function generate_system!(system::System, system_data::AbstractDict{Symbol,Any})
 
     # Load the assets
     load!(system, system_data[:assets])
+
     validate_unique_asset_ids(system)
+
+    # Load system-wide constraints
+    if haskey(system_data, :constraints)
+        @info(" -- Adding system-wide constraints")
+        check_and_convert_constraints!(system_data)
+        system.constraints = system_data[:constraints]
+    end
 
     @info("Done generating system. It took $(round(time() - start_time, digits=2)) seconds")
     return nothing