+++ src/model/system.jl
@@ -6,6 +6,7 @@ mutable struct System <: AbstractSystem
     assets::Vector{AbstractAsset}
     locations::Vector{Union{Node, Location}}
     input_data::Vector{Dict{Symbol,Any}}
+    constraints::Vector{AbstractTypeConstraint}
 end
 
 """
@@ -108,7 +109,8 @@ function empty_system(data_dirpath::String)
         Dict{Symbol,TimeData}(),
         [],
         [],
-        []
+        [],
+        Vector{AbstractTypeConstraint}()
     )
 end